Split backend/server.js into three router files according to their API = Mongo, Github, Cucumber/Selenium
